+++ packages/forms-transformer/lib/init/questions/input-source.js
@@ -4,17 +4,22 @@ module.exports = {
   name: 'inputSource',
   type: 'list',
   message: 'What is the source of your from definition?',
-  default: 'answerspace',
+  default: 'oneBlinkForms',
   choices: [
     {
-      name: 'AnswerSpace',
-      short: 'AnswerSpace',
-      value: 'answerspace'
+      name: 'OneBlink Forms',
+      short: 'OneBlink Forms',
+      value: 'oneBlinkForms'
     },
     {
       name: 'JSON Files',
       short: 'JSON Files',
       value: 'json_files'
+    },
+    {
+      name: 'AnswerSpace',
+      short: 'AnswerSpace',
+      value: 'answerspace'
     }
   ]
 }