+++ packages/forms-transformer/lib/utils/decide-on-normalisation-fns.js
@@ -5,10 +5,12 @@ const glob = require('glob')
 
 const bmpNormalise = require('../normalisations/from-bmp.js')
 const jsonNormalise = require('../normalisations/from-json-files.js')
+const oneBlinkFormsNormalise = require('../normalisations/from-one-blink-forms.js')
 
 const bmpCompile = (options) => () => bmpNormalise(options)
 const fileCompile = (options) => () => jsonNormalise(options.definitionSource)
 const getFileFromGlob = (options) => () => jsonNormalise(glob.sync(options.definitionSource))
+const oneBlinkFormsCompile = (options) => () => oneBlinkFormsNormalise(options)
 
 function decide (options) {
   switch (true) {
@@ -16,6 +18,8 @@ function decide (options) {
       return bmpCompile(options)
     case isGlob(options.definitionSource):
       return getFileFromGlob(options)
+    case Array.isArray(options.definitionSource):
+      return oneBlinkFormsCompile(options)
     default:
       return fileCompile(options)
   }