+++ packages/forms-transformer/lib/utils/get-auth-origin.js
@@ -1,0 +1,5 @@
+'use strict'
+
+module.exports = ((config) => {
+  return config.authorisationService && config.authorisationService.origin ? config.authorisationService.origin : 'https://auth-api.blinkm.io'
+})