+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - FT-141 | Store templates by form name in template service
+- FT-243 | Add OneBlink option for forms source
 
 ### Changed
 - FT-149 | default source path from `./bm-forms-src` to `./bm-forms/src`